fix error handler,optimize code
